requirements.txt
transformers
tqdm
nebula3-python
